+++ tools/generator/domain/src/main/java/is/codion/tools/generator/domain/DomainSource.java
@@ -26,6 +26,7 @@ import is.codion.framework.domain.entity.EntityDefinition;
 import is.codion.framework.domain.entity.EntityType;
 import is.codion.framework.domain.entity.KeyGenerator;
 import is.codion.framework.domain.entity.attribute.AttributeDefinition;
+import is.codion.framework.domain.entity.attribute.AuditColumnDefinition;
 import is.codion.framework.domain.entity.attribute.Column;
 import is.codion.framework.domain.entity.attribute.ColumnDefinition;
 import is.codion.framework.domain.entity.attribute.ForeignKey;
@@ -414,6 +415,11 @@ public final class DomainSource {
 						.append(interfaceName).append(".").append(column.name().toUpperCase()).append(".define()")
 						.append(LINE_SEPARATOR).append(TRIPLE_INDENT)
 						.append(".").append(definitionType(column, compositePrimaryKey));
+		if (auditColumn(column)) {
+			return builder.append(LINE_SEPARATOR).append(TRIPLE_INDENT)
+							.append(".caption(").append("\"").append(column.caption()).append("\")")
+							.toString();
+		}
 		if (!foreignKeyColumn && !column.primaryKey()) {
 			builder.append(LINE_SEPARATOR).append(TRIPLE_INDENT).append(".caption(").append("\"").append(column.caption()).append("\")");
 		}
@@ -456,8 +462,23 @@ public final class DomainSource {
 		if (column.primaryKey()) {
 			return compositePrimaryKey ? "primaryKey(" + column.primaryKeyIndex() + ")" : "primaryKey()";
 		}
+		if (auditColumn(column)) {
+			AuditColumnDefinition<?> auditColumnDefinition = (AuditColumnDefinition<?>) column;
+			switch (auditColumnDefinition.auditAction()) {
+				case INSERT:
+					return column.attribute().type().isString() ? "auditInsertUserColumn()" : "auditInsertTimeColumn()";
+				case UPDATE:
+					return column.attribute().type().isTemporal() ? "auditUpdateUserColumn()" : "auditUpdateTimeColumn()";
+				default:
+					throw new IllegalArgumentException("Uknown audit action: " + auditColumnDefinition.auditAction());
+			}
+		}
 
 		return "column()";
+	}
+
+	private static boolean auditColumn(ColumnDefinition<?> column) {
+		return column instanceof AuditColumnDefinition<?>;
 	}
 
 	private static String interfaceName(String tableName, boolean uppercase) {