+++ tools/generator/domain/src/test/resources/is/codion/tools/generator/domain/Petstore.java
@@ -18,6 +18,7 @@ import is.codion.petstore.domain.Petstore.TagItem;
 import is.codion.framework.domain.entity.EntityType;
 import is.codion.framework.domain.entity.attribute.Column;
 import is.codion.framework.domain.entity.attribute.ForeignKey;
+import java.time.LocalDateTime;
 
 public final class Petstore extends DomainModel {
 	public static final DomainType DOMAIN = domainType(Petstore.class);
@@ -71,7 +72,19 @@ public final class Petstore extends DomainModel {
 					.caption("Location"),
 				Address.IMAGE.define()
 					.column()
-					.caption("Image"))
+					.caption("Image"),
+				Address.INSERT_USER.define()
+					.auditInsertUserColumn()
+					.caption("Insert user"),
+				Address.INSERT_TIME.define()
+					.auditInsertTimeColumn()
+					.caption("Insert time"),
+				Address.UPDATE_USER.define()
+					.auditUpdateTimeColumn()
+					.caption("Update user"),
+				Address.UPDATE_TIME.define()
+					.auditUpdateUserColumn()
+					.caption("Update time"))
 			.keyGenerator(identity())
 			.caption("Address")
 			.build();
@@ -94,7 +107,19 @@ public final class Petstore extends DomainModel {
 				Category.IMAGE_URL.define()
 					.column()
 					.caption("Image url")
-					.maximumLength(55))
+					.maximumLength(55),
+				Category.INSERT_USER.define()
+					.auditInsertUserColumn()
+					.caption("Insert user"),
+				Category.INSERT_TIME.define()
+					.auditInsertTimeColumn()
+					.caption("Insert time"),
+				Category.UPDATE_USER.define()
+					.auditUpdateTimeColumn()
+					.caption("Update user"),
+				Category.UPDATE_TIME.define()
+					.auditUpdateUserColumn()
+					.caption("Update time"))
 			.keyGenerator(identity())
 			.caption("Category")
 			.build();
@@ -118,7 +143,19 @@ public final class Petstore extends DomainModel {
 					.column()
 					.caption("Email")
 					.nullable(false)
-					.maximumLength(24))
+					.maximumLength(24),
+				ContactInfo.INSERT_USER.define()
+					.auditInsertUserColumn()
+					.caption("Insert user"),
+				ContactInfo.INSERT_TIME.define()
+					.auditInsertTimeColumn()
+					.caption("Insert time"),
+				ContactInfo.UPDATE_USER.define()
+					.auditUpdateTimeColumn()
+					.caption("Update user"),
+				ContactInfo.UPDATE_TIME.define()
+					.auditUpdateUserColumn()
+					.caption("Update time"))
 			.keyGenerator(identity())
 			.caption("Contact info")
 			.build();
@@ -145,7 +182,19 @@ public final class Petstore extends DomainModel {
 					.column()
 					.caption("Tag")
 					.nullable(false)
-					.maximumLength(30))
+					.maximumLength(30),
+				Tag.INSERT_USER.define()
+					.auditInsertUserColumn()
+					.caption("Insert user"),
+				Tag.INSERT_TIME.define()
+					.auditInsertTimeColumn()
+					.caption("Insert time"),
+				Tag.UPDATE_USER.define()
+					.auditUpdateTimeColumn()
+					.caption("Update user"),
+				Tag.UPDATE_TIME.define()
+					.auditUpdateUserColumn()
+					.caption("Update time"))
 			.keyGenerator(identity())
 			.caption("Tag")
 			.build();
@@ -175,7 +224,19 @@ public final class Petstore extends DomainModel {
 				Product.IMAGE_URL.define()
 					.column()
 					.caption("Image url")
-					.maximumLength(55))
+					.maximumLength(55),
+				Product.INSERT_USER.define()
+					.auditInsertUserColumn()
+					.caption("Insert user"),
+				Product.INSERT_TIME.define()
+					.auditInsertTimeColumn()
+					.caption("Insert time"),
+				Product.UPDATE_USER.define()
+					.auditUpdateTimeColumn()
+					.caption("Update user"),
+				Product.UPDATE_TIME.define()
+					.auditUpdateUserColumn()
+					.caption("Update time"))
 			.keyGenerator(identity())
 			.caption("Product")
 			.description("The available products")
@@ -237,7 +298,19 @@ public final class Petstore extends DomainModel {
 					.column()
 					.caption("Disabled")
 					.nullable(false)
-					.columnHasDefaultValue(true))
+					.columnHasDefaultValue(true),
+				Item.INSERT_USER.define()
+					.auditInsertUserColumn()
+					.caption("Insert user"),
+				Item.INSERT_TIME.define()
+					.auditInsertTimeColumn()
+					.caption("Insert time"),
+				Item.UPDATE_USER.define()
+					.auditUpdateTimeColumn()
+					.caption("Update user"),
+				Item.UPDATE_TIME.define()
+					.auditUpdateUserColumn()
+					.caption("Update time"))
 			.keyGenerator(identity())
 			.caption("Item")
 			.build();
@@ -254,7 +327,19 @@ public final class Petstore extends DomainModel {
 					.primaryKey(1),
 				TagItem.ITEM_ID_FK.define()
 					.foreignKey()
-					.caption("Item"))
+					.caption("Item"),
+				TagItem.INSERT_USER.define()
+					.auditInsertUserColumn()
+					.caption("Insert user"),
+				TagItem.INSERT_TIME.define()
+					.auditInsertTimeColumn()
+					.caption("Insert time"),
+				TagItem.UPDATE_USER.define()
+					.auditUpdateTimeColumn()
+					.caption("Update user"),
+				TagItem.UPDATE_TIME.define()
+					.auditUpdateUserColumn()
+					.caption("Update time"))
 			.caption("Tag item")
 			.build();
 	}
@@ -272,6 +357,10 @@ public final class Petstore extends DomainModel {
 		Column<Double> LONGITUDE = TYPE.doubleColumn("longitude");
 		Column<Object> LOCATION = TYPE.column("location", Object.class);
 		Column<byte[]> IMAGE = TYPE.byteArrayColumn("image");
+		Column<String> INSERT_USER = TYPE.stringColumn("insert_user");
+		Column<LocalDateTime> INSERT_TIME = TYPE.localDateTimeColumn("insert_time");
+		Column<String> UPDATE_USER = TYPE.stringColumn("update_user");
+		Column<LocalDateTime> UPDATE_TIME = TYPE.localDateTimeColumn("update_time");
 	}
 
 	public interface Category {
@@ -281,6 +370,10 @@ public final class Petstore extends DomainModel {
 		Column<String> NAME = TYPE.stringColumn("name");
 		Column<String> DESCRIPTION = TYPE.stringColumn("description");
 		Column<String> IMAGE_URL = TYPE.stringColumn("image_url");
+		Column<String> INSERT_USER = TYPE.stringColumn("insert_user");
+		Column<LocalDateTime> INSERT_TIME = TYPE.localDateTimeColumn("insert_time");
+		Column<String> UPDATE_USER = TYPE.stringColumn("update_user");
+		Column<LocalDateTime> UPDATE_TIME = TYPE.localDateTimeColumn("update_time");
 	}
 
 	public interface ContactInfo {
@@ -290,6 +383,10 @@ public final class Petstore extends DomainModel {
 		Column<String> LAST_NAME = TYPE.stringColumn("last_name");
 		Column<String> FIRST_NAME = TYPE.stringColumn("first_name");
 		Column<String> EMAIL = TYPE.stringColumn("email");
+		Column<String> INSERT_USER = TYPE.stringColumn("insert_user");
+		Column<LocalDateTime> INSERT_TIME = TYPE.localDateTimeColumn("insert_time");
+		Column<String> UPDATE_USER = TYPE.stringColumn("update_user");
+		Column<LocalDateTime> UPDATE_TIME = TYPE.localDateTimeColumn("update_time");
 	}
 
 	public interface ItemTags {
@@ -304,6 +401,10 @@ public final class Petstore extends DomainModel {
 
 		Column<Integer> TAG_ID = TYPE.integerColumn("tag_id");
 		Column<String> TAG = TYPE.stringColumn("tag");
+		Column<String> INSERT_USER = TYPE.stringColumn("insert_user");
+		Column<LocalDateTime> INSERT_TIME = TYPE.localDateTimeColumn("insert_time");
+		Column<String> UPDATE_USER = TYPE.stringColumn("update_user");
+		Column<LocalDateTime> UPDATE_TIME = TYPE.localDateTimeColumn("update_time");
 	}
 
 	public interface Product {
@@ -314,6 +415,10 @@ public final class Petstore extends DomainModel {
 		Column<String> NAME = TYPE.stringColumn("name");
 		Column<String> DESCRIPTION = TYPE.stringColumn("description");
 		Column<String> IMAGE_URL = TYPE.stringColumn("image_url");
+		Column<String> INSERT_USER = TYPE.stringColumn("insert_user");
+		Column<LocalDateTime> INSERT_TIME = TYPE.localDateTimeColumn("insert_time");
+		Column<String> UPDATE_USER = TYPE.stringColumn("update_user");
+		Column<LocalDateTime> UPDATE_TIME = TYPE.localDateTimeColumn("update_time");
 
 		ForeignKey CATEGORYID_FK = TYPE.foreignKey("categoryid_fk", CATEGORYID, Category.CATEGORY_ID);
 	}
@@ -333,6 +438,10 @@ public final class Petstore extends DomainModel {
 		Column<Integer> TOTAL_SCORE = TYPE.integerColumn("total_score");
 		Column<Integer> NUMBER_OF_VOTES = TYPE.integerColumn("number_of_votes");
 		Column<Integer> DISABLED = TYPE.integerColumn("disabled");
+		Column<String> INSERT_USER = TYPE.stringColumn("insert_user");
+		Column<LocalDateTime> INSERT_TIME = TYPE.localDateTimeColumn("insert_time");
+		Column<String> UPDATE_USER = TYPE.stringColumn("update_user");
+		Column<LocalDateTime> UPDATE_TIME = TYPE.localDateTimeColumn("update_time");
 
 		ForeignKey PRODUCT_ID_FK = TYPE.foreignKey("product_id_fk", PRODUCT_ID, Product.PRODUCT_ID);
 		ForeignKey ADDRESS_ID_FK = TYPE.foreignKey("address_id_fk", ADDRESS_ID, Address.ADDRESS_ID);
@@ -344,6 +453,10 @@ public final class Petstore extends DomainModel {
 
 		Column<Integer> TAG_ID = TYPE.integerColumn("tag_id");
 		Column<Integer> ITEM_ID = TYPE.integerColumn("item_id");
+		Column<String> INSERT_USER = TYPE.stringColumn("insert_user");
+		Column<LocalDateTime> INSERT_TIME = TYPE.localDateTimeColumn("insert_time");
+		Column<String> UPDATE_USER = TYPE.stringColumn("update_user");
+		Column<LocalDateTime> UPDATE_TIME = TYPE.localDateTimeColumn("update_time");
 
 		ForeignKey TAG_ID_FK = TYPE.foreignKey("tag_id_fk", TAG_ID, Tag.TAG_ID);
 		ForeignKey ITEM_ID_FK = TYPE.foreignKey("item_id_fk", ITEM_ID, Item.ITEM_ID);