+++ tools/generator/domain/src/test/resources/is/codion/tools/generator/domain/PetstoreAPI.java
@@ -6,6 +6,7 @@ import is.codion.framework.domain.DomainType;
 import is.codion.framework.domain.entity.EntityType;
 import is.codion.framework.domain.entity.attribute.Column;
 import is.codion.framework.domain.entity.attribute.ForeignKey;
+import java.time.LocalDateTime;
 
 public interface Petstore {
 	DomainType DOMAIN = domainType(Petstore.class);
@@ -23,6 +24,10 @@ public interface Petstore {
 		Column<Double> LONGITUDE = TYPE.doubleColumn("longitude");
 		Column<Object> LOCATION = TYPE.column("location", Object.class);
 		Column<byte[]> IMAGE = TYPE.byteArrayColumn("image");
+		Column<String> INSERT_USER = TYPE.stringColumn("insert_user");
+		Column<LocalDateTime> INSERT_TIME = TYPE.localDateTimeColumn("insert_time");
+		Column<String> UPDATE_USER = TYPE.stringColumn("update_user");
+		Column<LocalDateTime> UPDATE_TIME = TYPE.localDateTimeColumn("update_time");
 	}
 
 	interface Category {
@@ -32,6 +37,10 @@ public interface Petstore {
 		Column<String> NAME = TYPE.stringColumn("name");
 		Column<String> DESCRIPTION = TYPE.stringColumn("description");
 		Column<String> IMAGE_URL = TYPE.stringColumn("image_url");
+		Column<String> INSERT_USER = TYPE.stringColumn("insert_user");
+		Column<LocalDateTime> INSERT_TIME = TYPE.localDateTimeColumn("insert_time");
+		Column<String> UPDATE_USER = TYPE.stringColumn("update_user");
+		Column<LocalDateTime> UPDATE_TIME = TYPE.localDateTimeColumn("update_time");
 	}
 
 	interface ContactInfo {
@@ -41,6 +50,10 @@ public interface Petstore {
 		Column<String> LAST_NAME = TYPE.stringColumn("last_name");
 		Column<String> FIRST_NAME = TYPE.stringColumn("first_name");
 		Column<String> EMAIL = TYPE.stringColumn("email");
+		Column<String> INSERT_USER = TYPE.stringColumn("insert_user");
+		Column<LocalDateTime> INSERT_TIME = TYPE.localDateTimeColumn("insert_time");
+		Column<String> UPDATE_USER = TYPE.stringColumn("update_user");
+		Column<LocalDateTime> UPDATE_TIME = TYPE.localDateTimeColumn("update_time");
 	}
 
 	interface ItemTags {
@@ -55,6 +68,10 @@ public interface Petstore {
 
 		Column<Integer> TAG_ID = TYPE.integerColumn("tag_id");
 		Column<String> TAG = TYPE.stringColumn("tag");
+		Column<String> INSERT_USER = TYPE.stringColumn("insert_user");
+		Column<LocalDateTime> INSERT_TIME = TYPE.localDateTimeColumn("insert_time");
+		Column<String> UPDATE_USER = TYPE.stringColumn("update_user");
+		Column<LocalDateTime> UPDATE_TIME = TYPE.localDateTimeColumn("update_time");
 	}
 
 	interface Product {
@@ -65,6 +82,10 @@ public interface Petstore {
 		Column<String> NAME = TYPE.stringColumn("name");
 		Column<String> DESCRIPTION = TYPE.stringColumn("description");
 		Column<String> IMAGE_URL = TYPE.stringColumn("image_url");
+		Column<String> INSERT_USER = TYPE.stringColumn("insert_user");
+		Column<LocalDateTime> INSERT_TIME = TYPE.localDateTimeColumn("insert_time");
+		Column<String> UPDATE_USER = TYPE.stringColumn("update_user");
+		Column<LocalDateTime> UPDATE_TIME = TYPE.localDateTimeColumn("update_time");
 
 		ForeignKey CATEGORYID_FK = TYPE.foreignKey("categoryid_fk", CATEGORYID, Category.CATEGORY_ID);
 	}
@@ -84,6 +105,10 @@ public interface Petstore {
 		Column<Integer> TOTAL_SCORE = TYPE.integerColumn("total_score");
 		Column<Integer> NUMBER_OF_VOTES = TYPE.integerColumn("number_of_votes");
 		Column<Integer> DISABLED = TYPE.integerColumn("disabled");
+		Column<String> INSERT_USER = TYPE.stringColumn("insert_user");
+		Column<LocalDateTime> INSERT_TIME = TYPE.localDateTimeColumn("insert_time");
+		Column<String> UPDATE_USER = TYPE.stringColumn("update_user");
+		Column<LocalDateTime> UPDATE_TIME = TYPE.localDateTimeColumn("update_time");
 
 		ForeignKey PRODUCT_ID_FK = TYPE.foreignKey("product_id_fk", PRODUCT_ID, Product.PRODUCT_ID);
 		ForeignKey ADDRESS_ID_FK = TYPE.foreignKey("address_id_fk", ADDRESS_ID, Address.ADDRESS_ID);
@@ -95,6 +120,10 @@ public interface Petstore {
 
 		Column<Integer> TAG_ID = TYPE.integerColumn("tag_id");
 		Column<Integer> ITEM_ID = TYPE.integerColumn("item_id");
+		Column<String> INSERT_USER = TYPE.stringColumn("insert_user");
+		Column<LocalDateTime> INSERT_TIME = TYPE.localDateTimeColumn("insert_time");
+		Column<String> UPDATE_USER = TYPE.stringColumn("update_user");
+		Column<LocalDateTime> UPDATE_TIME = TYPE.localDateTimeColumn("update_time");
 
 		ForeignKey TAG_ID_FK = TYPE.foreignKey("tag_id_fk", TAG_ID, Tag.TAG_ID);
 		ForeignKey ITEM_ID_FK = TYPE.foreignKey("item_id_fk", ITEM_ID, Item.ITEM_ID);